move the new components
